chore(reviews): rename google-customer-reviews.js to match its export name
File exported GoogleCustomerReviewsSettings but was named without the
-settings suffix, unlike sibling settings components (e.g.
setup-enhanced-conversions.js). Renames the component, its co-located
test, and its scss file to match, updating the barrel index.js export.

GOOWOO-869
